+++ pierky/arouteserver/builder.py
@@ -31,6 +31,7 @@ from .enrichers.irrdb import IRRDBConfigEnricher_OriginASNs, \
                              IRRDBConfigEnricher_Prefixes
 from .enrichers.pdb_as_set import PeeringDBConfigEnricher_ASSet
 from .enrichers.pdb_max_prefix import PeeringDBConfigEnricher_MaxPrefix
+from .enrichers.rpki_roas import RPKIROAsEnricher
 from .enrichers.rtt import RTTGetterConfigEnricher
 from .errors import MissingDirError, MissingFileError, BuilderError, \
                     ARouteServerError, MissingArgumentError, \
@@ -410,6 +411,9 @@ class ConfigBuilder(object):
         # { "<as_set_bundle_id>": <AS_SET_Bundle_Proxy>, ... }
         self.irrdb_info = None
 
+        # { "<origin_asn>": [{"prefix": "a/b", "max_len": c}] }
+        self.rpki_roas_as_route_objects = {}
+
         # Validation
 
         if self.local_files:
@@ -476,15 +480,27 @@ class ConfigBuilder(object):
         # Enrichers
         # Order matters: AS-SET from PeeringDB must be run first
         # in order to acquire missing AS-SETs that are processed
-        # later by IRRDB enrichers.
+        # later by IRRDB enrichers. RPKI ROAs are processed only
+        # for those origin ASNs that have been gathered from AS-SETs.
+        irrdb_cfg = self.cfg_general["filtering"]["irrdb"]
         used_enricher_classes = []
-        if self.cfg_general["filtering"]["irrdb"]["peering_db"]:
+
+        if irrdb_cfg["peering_db"]:
             used_enricher_classes += [PeeringDBConfigEnricher_ASSet]
+
         used_enricher_classes += [IRRDBConfigEnricher_OriginASNs,
                                   IRRDBConfigEnricher_Prefixes,
                                   PeeringDBConfigEnricher_MaxPrefix]
+
         if self.cfg_general.rtt_based_functions_are_used:
             used_enricher_classes.append(RTTGetterConfigEnricher)
+
+        if irrdb_cfg["enforce_origin_in_as_set"] and \
+            irrdb_cfg["enforce_prefix_in_as_set"] and \
+            irrdb_cfg["use_rpki_roas_as_route_objects"]["enabled"] and \
+            irrdb_cfg["use_rpki_roas_as_route_objects"]["source"] == \
+                "ripe-rpki-validator-cache":
+            used_enricher_classes.append(RPKIROAsEnricher)
 
         for enricher_class in used_enricher_classes:
             enricher = enricher_class(self, threads=self.threads)
@@ -526,6 +542,7 @@ class ConfigBuilder(object):
         self.data["clients"] = self.cfg_clients
         self.data["asns"] = self.cfg_asns
         self.data["irrdb_info"] = self.irrdb_info
+        self.data["rpki_roas_as_route_objects"] = self.rpki_roas_as_route_objects
         self.data["roas"] = self.cfg_roas
         self.data["live_tests"] = self.live_tests
         self.data["rtt_based_functions_are_used"] = \
@@ -701,7 +718,8 @@ class OpenBGPDConfigBuilder(ConfigBuilder):
                         "add_path", "max_prefix_action",
                         "blackhole_filtering_rewrite_ipv6_nh",
                         "large_communities", "extended_communities",
-                        "graceful_shutdown", "rfc1997_wellknown_communities"]
+                        "graceful_shutdown", "rfc1997_wellknown_communities",
+                        "rpki_roas_as_route_objects_internal_community"]
 
     def _include_local_file(self, local_file_id):
         return 'include "{}"\n\n'.format(
@@ -946,6 +964,30 @@ class OpenBGPDConfigBuilder(ConfigBuilder):
                 ):
                     res = False
 
+        use_rpki_roas_as_route_objects_cfg = \
+            self.cfg_general["filtering"]["irrdb"]["use_rpki_roas_as_route_objects"]
+        if use_rpki_roas_as_route_objects_cfg["enabled"]:
+            if use_rpki_roas_as_route_objects_cfg["source"] != "ripe-rpki-validator-cache":
+                res = False
+                logging.error(
+                    "For OpenBGPD only the 'ripe-rpki-validator-cache'"
+                    "value is allowed for the "
+                    "'use_rpki_roas_as_route_objects.source' option."
+                )
+
+            for comm_name in ConfigParserGeneral.COMMUNITIES_SCHEMA:
+                comm = self.cfg_general["communities"][comm_name]
+                if comm["ext"] and comm["ext"] == "ro:65535:1":
+                    if not self.process_bgpspeaker_specific_compatibility_issue(
+                        "rpki_roas_as_route_objects_internal_community",
+                        "When the 'use_rpki_roas_as_route_objects' option "
+                        "is set, the ro:65535:1 extended community must "
+                        "be reserved for internal purposes. "
+                        "A collision has been detected with the following "
+                        "community: {}".format(comm_name)
+                    ):
+                        res = False
+
         return res
 
     def enrich_j2_environment(self, env):
@@ -995,5 +1037,12 @@ class TemplateContextDumper(ConfigBuilder):
                 lst.append(bundle.to_dict())
             return lst
 
+        def parse_rpki_roas_as_route_objects(rpki_roas_as_route_objects):
+            res = {}
+            for origin_asn in rpki_roas_as_route_objects:
+                res[origin_asn] = list(rpki_roas_as_route_objects[origin_asn].roas)
+            return res
+
         env.filters["to_yaml"] = to_yaml
         env.filters["parse_irrdb_info"] = parse_irrdb_info
+        env.filters["parse_rpki_roas_as_route_objects"] = parse_rpki_roas_as_route_objects