+++ pierky/arouteserver/errors.py
@@ -100,6 +100,9 @@ class IXFDBSchemaError(IXFDBError):
                             "the one recognized by this version of the "
                             "program, or that it contains errors.")
 
+class RPKIValidatorCacheError(ARouteServerError):
+    pass
+
 class BuilderError(ARouteServerError):
     pass